% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frailty_DEIE.r
\name{Frailty.DEIE}
\alias{Frailty.DEIE}
\title{Estimating the parameters of the frailty model}
\usage{
Frailty.DEIE(data, P.time, int_theta, tol, step)
}
\description{
Estimating the parameters of the frailty model
}
\examples{
Frailty.DEIE()
}
\keyword{Frailty.DEIE}
