% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_frailty_free_theta.r
\name{meta.gen}
\alias{meta.gen}
\title{generating the data with 1/2 bivariate exposure from the frailty model by  three exponential distributions}
\usage{
meta.gen(
  n,
  theta_0,
  theta_1,
  L1,
  L2,
  L3,
  b01,
  b02,
  b03,
  cc = 2,
  dd = "uniform"
)
}
\arguments{
\item{n}{sample size (even)}

\item{theta_0}{theta for Z=1}

\item{theta_1}{theta for Z=2}

\item{L1}{lambda for genarate T1}

\item{L2}{lambda for generate T2 without given T1}

\item{L3}{lambda for generate T2 given T1}

\item{b01}{effect from Z to T1}

\item{b02}{effect from Z to T2}

\item{b03}{effect from T1 to T2}

\item{cc}{parameter for generating the censoring time, the regulator censoring rate}

\item{dd}{set "uniform" for U(0,cc); set weibull for weibull(shape=5,scale=cc)}

\item{output}{X1,X2,Z and D are observed mediated, terminal event times exposure and censoring index (1/0 for failure and censored)}
}
\description{
generating the data with 1/2 bivariate exposure from the frailty model by  three exponential distributions
}
\examples{
meta.gen(500,theta_0=1,theta_1=0.5,L1=1,L2=1,L3=1,b01=0.5,b02=0,b03=1,cc=2,dd="uniform")
}
\keyword{meta.gen}
