% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLA.r
\name{pLA}
\alias{pLA}
\title{predicting the value for one to one vector}
\usage{
pLA(yy, tt, LL)
}
\arguments{
\item{yy}{any time you want to interpolate}

\item{tt}{time or x value with same length of LL, must >0}

\item{LL}{value as f(tt)}
}
\description{
predicting the value for one to one vector
}
\examples{
x=seq(0,0.5,by=0.01)
LL=pnorm(x,0,1)
pLA(c(0,0.1,0.25,0.01,3),x,LL)
}
\keyword{pLA}
